CTRL: break out control manager and valve into separate crates
